requirements.txt
django == 2.2
djangorestframework
postgres
psycopg2
pillow

